worker agent pre-check
